HomieGramBackEnd/requirements.txt
asgiref==3.8.1
boto3==1.35.34
botocore==1.35.34
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Django==5.1.1
django-rest-knox==5.0.2
djangorestframework==3.15.2
dynaconf==3.2.6
hvac==2.3.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
knox==0.1.14
loguru==0.7.2
MarkupSafe==2.1.5
pillow==10.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.2
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
validators==0.34.0
